cmake: Guard h5md linking. Moved Vector to utils folder
