+++ src/core/ComFixed.hpp
@@ -19,7 +19,7 @@
 #ifndef CORE_COMFIXED_HPP
 #define CORE_COMFIXED_HPP
 
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 #include "utils/keys.hpp"
 
 #include <boost/mpi/collectives/all_reduce.hpp>
