+++ src/core/cluster_analysis/Cluster.cpp
@@ -22,7 +22,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #ifdef GSL
 #include "gsl/gsl_fit.h"
 #endif
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 #include <vector>
 
 #include "Cluster.hpp"
