+++ src/core/cluster_analysis/Cluster.hpp
@@ -19,7 +19,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef CLUSTER_ANALYSIS_CLUSTER_HPP
 #define CLUSTER_ANALYSIS_CLUSTER_HPP
 
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 #include <map>
 #include <string>
 #include <valarray>
