+++ src/core/electrostatics_magnetostatics/icc.hpp
@@ -57,7 +57,7 @@
 
 #if defined(ELECTROSTATICS)
 
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 
 /* iccp3m data structures*/
 struct iccp3m_struct {
