+++ src/core/field_coupling/couplings/Viscous.hpp
@@ -19,7 +19,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef CORE_CONSTRAINTS_DETAIL_VISCOUS_HPP
 #define CORE_CONSTRAINTS_DETAIL_VISCOUS_HPP
 
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 
 namespace FieldCoupling {
 namespace Coupling {
