+++ src/core/field_coupling/fields/Constant.hpp
@@ -19,7 +19,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef CORE_EXTERNAL_FIELD_FIELDS_CONSTANT_HPP
 #define CORE_EXTERNAL_FIELD_FIELDS_CONSTANT_HPP
 
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 #include "gradient_type.hpp"
 
 namespace FieldCoupling {
