+++ src/core/observables/CylindricalProfileObservable.hpp
@@ -22,7 +22,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include <cmath>
 
 #include "Observable.hpp"
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 
 namespace Observables {
