+++ src/core/particle_data.hpp
@@ -25,7 +25,7 @@
  *  see \ref particle_data.cpp "particle_data.cpp"
  */
 
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 #include "config.hpp"
 #include "utils.hpp"
