+++ src/core/shapes/HollowCone.hpp
@@ -23,7 +23,7 @@
 #define __HOLLOWCONE_HPP
 
 #include "Shape.hpp"
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 
 namespace Shapes {
