+++ src/core/shapes/Rhomboid.hpp
@@ -23,7 +23,7 @@
 #define __RHOMBOID_HPP
 
 #include "Shape.hpp"
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 
 namespace Shapes {
 class Rhomboid : public Shape {
