+++ src/core/shapes/Stomatocyte.hpp
@@ -23,7 +23,7 @@
 #define __STOMATOCYTE_HPP
 
 #include "Shape.hpp"
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 
 namespace Shapes {
 class Stomatocyte : public Shape {
