+++ src/core/shapes/Wall.hpp
@@ -23,7 +23,7 @@
 #define SHAPES_WALL_HPP
 
 #include "Shape.hpp"
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 
 namespace Shapes {
 class Wall : public Shape {
