+++ src/core/utils/interpolation/detail/ll_and_dist.hpp
@@ -19,7 +19,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef UTILS_INTERPOLATION_DETAIL_LL_AND_DIST_HPP
 #define UTILS_INTERPOLATION_DETAIL_LL_AND_DIST_HPP
 
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 
 #include <array>
 #include <cmath>
