+++ src/core/utils/vec_rotate.hpp
@@ -1,7 +1,7 @@
 #ifndef UTILS_VEC_ROTATE_HPP
 #define UTILS_VEC_ROTATE_HPP
 
-#include "Vector.hpp"
+#include "utils/Vector.hpp"
 #include "utils.hpp"
 #include "utils/math/sqr.hpp"
