+++ src/script_interface/CMakeLists.txt
@@ -36,5 +36,8 @@ endif()
 add_library(EspressoScriptInterface SHARED ${EspressoScriptInterface_SRC})
 install(TARGETS EspressoScriptInterface LIBRARY DESTINATION ${PYTHON_INSTDIR}/espressomd)
 set_target_properties(EspressoScriptInterface PROPERTIES SOVERSION ${SOVERSION})
-target_link_libraries(EspressoScriptInterface PRIVATE EspressoConfig PUBLIC EspressoCore Shapes H5mdCore LBBoundaries cluster_analysis mpiio pdbreader)
+target_link_libraries(EspressoScriptInterface PRIVATE EspressoConfig PUBLIC EspressoCore Shapes LBBoundaries cluster_analysis mpiio pdbreader)
+if(HDF5_FOUND)
+ target_link_libraries(EspressoScriptInterface PRIVATE H5mdCore)
+endif()
 target_include_directories(EspressoScriptInterface PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
